Cleaning up content page and hiding CTA for closed state
