+++ resources/assets/components/App.js
@@ -11,7 +11,7 @@ import ChromeContainer from '../containers/ChromeContainer';
 import FeedContainer from '../containers/FeedContainer';
 import { ActionPageContainer } from './ActionPage';
 import { BlockContainer } from './Block';
-import ContentPageContainer from '../containers/ContentPageContainer';
+import { ContentPageContainer } from './ContentPage';
 import NotFound from './NotFound';
 
 const wrap = (Container, Component) => props => (
